intial commit for spring 2016
mainly contains s core classes but also kim and kens class. going to be a ough semester as its 22 credits. will not be commiting everyday like last year as that is pretty time consuming but will be updating peroidcally. main use of this is as a backup for the little code on the go machine